requirements.txt
Flask==1.1.2
google-cloud-pubsub==1.4.2
google-cloud-storage==1.28.0
GoogleAppEngineCloudStorageClient==1.9.22.1
google-cloud-datastore==1.12.0
google-auth==1.14.1
requests==2.23.0